Move test to other folder, fix wrong code
