+++ test/decoder_test.h
@@ -1,0 +1,8 @@
+#ifndef DECOTEST_H
+#define DECOTEST_H
+#include <assert.h>
+#include <stdio.h>
+
+void decoding_tests();
+
+#endif
